requirements.txt
azure-monitor-opentelemetry
django==4.1
psycopg2-binary
python-dotenv
whitenoise
django-sslserver
azure-storage-blob
azure-identity
opentelemetry-api
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation
opentelemetry-instrumentation-django
opentelemetry-instrumentation-logging
opentelemetry-instrumentation-psycopg2
psutil
pyodbc